requirements.txt
six
appdirs==1.4.4
arabic-reshaper==3.0.0
asgiref==3.2.10
asn1crypto==1.5.1
autopep8==2.0.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
cssselect2==0.7.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.6
django-crum==0.7.7
django-heroku==0.3.1
django-ipware==3.0.2
django-widget-tweaks==1.4.8
filelock==3.0.12
flake8==2.3.0
future==0.18.2
gunicorn==20.1.0
html5lib
idna==3.4
lxml==4.9.3
mccabe==0.6.1
oscrypto==1.3.0
pep8==1.7.1
Pillow==10.0.0
psycopg2==2.9.7
pycodestyle==2.11.0
pycparser==2.21
pyflakes==2.2.0
pyHanko==0.20.0
pyhanko-certvalidator==0.23.0
pypdf==3.16.0
PyPDF2==1.26.0
pypng==0.20220715.0
python-bidi==0.4.2
python-decouple==3.4
pytz==2021.1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
six==1.15.0
sqlparse==0.4.1
svglib==1.5.1
tinycss2==1.2.1
toml==0.10.2
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.2
urllib3==2.0.4
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf
